+++ sdk/confidentialledger/Azure.Security.ConfidentialLedger/src/ConfidentialLedgerCertificateClient.cs
@@ -7,10 +7,11 @@ using System.Text.Json;
 using Azure.Core;
 using Azure.Core.Pipeline;
 
-namespace Azure.Security.ConfidentialLedger
+namespace Azure.Security.ConfidentialLedger.Certificate
 {
     [CodeGenSuppress("ConfidentialLedgerCertificateClient", typeof(Uri), typeof(TokenCredential), typeof(ConfidentialLedgerClientOptions))]
     [CodeGenSuppress("ConfidentialLedgerCertificateClient", typeof(Uri), typeof(TokenCredential))]
+    [CodeGenClient("ConfidentialLedgerCertificateClient")]
     public partial class ConfidentialLedgerCertificateClient
     {
         /// <summary> Initializes a new instance of ConfidentialLedgerCertificateClient. </summary>